move equalnullsafe to compare functions file
